+++ CHANGELOG.md
@@ -3,13 +3,15 @@
 
 -------------------------------------------------------------------------------------------------------------
 
-## 5.3.6 (2020-05-17)
+## 5.3.6 (2020-05-19)
 
 ### 新特性
 * 【core   】     NumberConverter Long类型增加日期转换（pr#872@Github）
 * 【all    】     StrUtil and SymmetricCrypto注释修正（pr#873@Github）
 * 【core   】     CsvReader支持返回Bean（issue#869@Github）
 * 【core   】     Snowflake循环等待下一个时间时避免长时间循环，加入对时钟倒退的判断（pr#874@Github）
+* 【extra  】     新增 QRCode base64 编码形式返回（pr#878@Github）
+* 【core   】     ImgUtil增加toBase64DateUri，URLUtil增加getDataUri方法
 
 ### Bug修复